Modified the AI maps.

+++ BattleCity/AiMap.h
@@ -13,7 +13,7 @@ public:
 private:
 	size_t m_width;
 	uint16_t m_height;
-	std::vector<std::vector<int>> m_map;
+	std::vector<std::vector<uint8_t>> m_map;
 
 };
